predefined roles,  access list as string

+++ classes/User/model/Role.php
@@ -105,21 +105,76 @@ class User_Role extends DBx_Table_Row
                 ->where( 'ucal_role_id = ?', $this->getId() );
         return User_AccessList::Table()->fetchAll($select);
     }
+    /**
+     * Getting Roles Access list as string
+     * Each group 
+     * @return string
+     */
+    public function getAccessListString( $strRoleSeparator = ', ', $strGroupSeparator = '<br />' )
+    {
+        if ( $this->isPredefined() ) {
+            return Lang_Hash::get( 'Predefined Role' );
+        }
+        $arrLines = array();
+        
+        $arrAccessList = array();
+        foreach( $this->getAccessList() as $objAccessRecord ) 
+            $arrAccessList[ $objAccessRecord->getResourceId() ] = $objAccessRecord->getResourceId();
+        
+        $config = App_Application::getInstance()->getConfig()->user->resource; 
+        foreach ( $config as $strGroup => $arrResource ) {
+            $arrGroup = array();
+            foreach ( $arrResource as $strName => $nResourceId ) {
+                 if ( isset( $arrAccessList[ $nResourceId ])  ) 
+                     $arrGroup[] = $strName;
+            }
+            if ( count( $arrGroup ) > 0 ) {
+                $arrLines[] = $strGroup .': '.implode( $strRoleSeparator, $arrGroup );
+            }
+        }
+        return implode ( $strGroupSeparator, $arrLines );
+    }
+    /**
+     * @param integer $nResourceId
+     * @return boolean
+     */
+    public function canAccessByResourceId( $nResourceId )
+    {
+        $arrAccessList = array();
+        foreach( $this->getAccessList() as $objAccessRecord ) 
+            $arrAccessList[ $objAccessRecord->getResourceId() ] = $objAccessRecord->getResourceId();
+        
+        return isset( $arrAccessList[ $nResourceId ] );
+    }
+    
+    
+    public function isPredefined()
+    {
+        // TODO: check configs for prefefined roles, return Administrator - if not found
+        return ( $this->getName() == 'Administrator' );
+    }
 
     /**
      * @return void
      */
     public function _delete()
     {
-        //delete all user-roles with this role
-        $selectRoles = User_UserRole::Table()->select()->where('ucur_role_id = ?', $this->getId() );
-        foreach ( User_UserRole::Table()->fetchAll( $selectRoles ) as $objUserRole) {
-            $objUserRole->delete();
+        if ( !$this->isPredefined() ) {
+            //delete all user-roles with this role
+            $selectRoles = User_UserRole::Table()->select()->where('ucur_role_id = ?', $this->getId() );
+            foreach ( User_UserRole::Table()->fetchAll( $selectRoles ) as $objUserRole) {
+                $objUserRole->delete();
+            }
+            //delete all access list for this resource
+            $selectList = User_AccessList::Table()->select()->where('ucal_role_id = ? ', $this->getId() );
+            foreach ( User_AccessList::Table()->fetchAll( $selectList ) as $objAccessList) {
+                $objAccessList->delete();
+            }        
         }
-        //delete all access list for this resource
-        $selectList = User_AccessList::Table()->select()->where('ucal_role_id = ? ', $this->getId() );
-        foreach ( User_AccessList::Table()->fetchAll( $selectList ) as $objAccessList) {
-            $objAccessList->delete();
-        }        
+    }
+    public function delete() {
+        if ( !$this->isPredefined() ) {
+            parent::delete();
+        }
     }
 }
